requirements.txt
pyaml>=3.11
networkx>=1.8.1
mathutils>=2.70
